Support different types of input data
The input that are parsed by the SaltStack outputters
can be of different types: int, string, list, dict.
For instance a 'cmd.run' will return a string of data,
a 'service.reload' command will give as output a boolean
value, and a 'state.apply' command will return a JSON object.

Modify the code to allow non JSON inputs.

Signed-off-by: Davide Madrisan <davide.madrisan@gmail.com>